@@ -2,7 +2,7 @@
   "private": true,
   "appVersion": "Sonic Rainboom",
   "version": "1.0.0",
-  "updatorVersion": 30,
+  "updatorVersion": 31,
   "sideEffects": false,
   "scripts": {
     "lc": "yarn lcd",
@@ -22,7 +22,7 @@
     "macos": "yarn electron-packager dist/release Alicorn --overwrite --platform=darwin --arch=x64 --icon ./resources/build/icon --out out",
     "linux": "yarn electron-packager dist/release Alicorn --overwrite --platform=linux --arch=x64 --icon ./resources/build/icon --out out&&node makesh.js",
     "clean": "node clean.js",
-    "launch-dev": "cd dist&&export ALICORN_REACT_DEVTOOLS=1&&npx electron .",
+    "launch-dev": "cd dist&&npx electron .",
     "launch-release": "cd dist&&cd release&&npx electron .",
     "make": "yarn clean&&yarn release-full&&yarn windows&&yarn linux&&yarn compress",
     "github-unix": "yarn clean&&yarn release-full&&yarn linux&&yarn compress",
@@ -34,38 +34,40 @@
     "@types/fs-extra": "^9.0.13",
     "@types/js-levenshtein": "^1.1.0",
     "@types/object-hash": "^2.2.1",
-    "@types/react": "^17.0.34",
+    "@types/react": "^17.0.35",
     "@types/react-dom": "^17.0.11",
     "@types/react-router-dom": "^5.3.2",
     "@types/semver": "^7.3.9",
     "@types/sha": "^3.0.1",
     "@types/throttle-debounce": "^2.1.0",
     "@types/tunnel": "^0.0.3",
-    "@types/uuid": "^8.3.1",
+    "@types/uuid": "^8.3.3",
     "@types/ws": "^8.2.0",
-    "@typescript-eslint/eslint-plugin": "^5.3.0",
-    "@typescript-eslint/parser": "^5.3.0",
-    "copy-webpack-plugin": "^9.0.1",
-    "electron": "^15.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "copy-webpack-plugin": "^10.0.0",
+    "electron": "^16.0.1",
     "electron-packager": "^15.4.0",
-    "eslint": "7.32.0",
+    "eslint": "8.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "prettier": "^2.4.1",
     "react-devtools": "^4.21.0",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.4.4",
-    "webpack": "^5.61.0",
+    "typescript": "^4.5.2",
+    "webpack": "^5.64.1",
     "webpack-cli": "^4.9.1",
     "yarn": "^1.22.17"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "ajv": "^8.6.3",
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.1.1",
+    "@mui/material": "^5.1.1",
+    "@mui/styles": "^5.1.1",
+    "ajv": "^8.8.1",
     "cheerio": "^1.0.0-rc.10",
     "compressing": "^1.5.1",
     "copy-to-clipboard": "^3.3.1",
