@@ -1,3 +1,4 @@
+import { DeleteForever, EventBusy, Sync } from "@mui/icons-material";
 import {
   Box,
   Card,
@@ -7,8 +8,7 @@ import {
   LinearProgress,
   Tooltip,
   Typography,
-} from "@material-ui/core";
-import { DeleteForever, EventBusy, Sync } from "@material-ui/icons";
+} from "@mui/material";
 import { remove } from "fs-extra";
 import objectHash from "object-hash";
 import React, { useEffect, useRef, useState } from "react";
@@ -17,6 +17,7 @@ import { scanCoresInAllMountedContainers } from "../modules/container/ContainerS
 import { getContainer } from "../modules/container/ContainerUtil";
 import { loadProfile } from "../modules/profile/ProfileLoader";
 import { whatProfile } from "../modules/profile/WhatProfile";
+import { setDirtyProfile } from "../modules/readyboom/PrepareProfile";
 import { jumpTo, triggerSetPage } from "./GoTo";
 import { Icons } from "./Icons";
 import { YNDialog2 } from "./OperatingHint";
@@ -119,7 +120,7 @@ function CoresDisplay(props: { server?: string }): JSX.Element {
 
   return (
     <>
-      <Box style={{ textAlign: "right" }}>
+      <Box sx={{ textAlign: "right" }}>
         <Tooltip
           title={
             <Typography className={"smtxt"}>{tr("CoreInfo.Reload")}</Typography>
@@ -141,7 +142,7 @@ function CoresDisplay(props: { server?: string }): JSX.Element {
         <>
           <LinearProgress color={"secondary"} />
           <Typography
-            style={{ fontSize: "medium", color: "#ff8400" }}
+            sx={{ fontSize: "medium", color: "#ff8400" }}
             gutterBottom
           >
             {tr("CoreInfo.StillLoading")}
@@ -183,6 +184,7 @@ function SingleCoreDisplay(props: {
   return (
     <>
       <Card
+        raised={true}
         onMouseOver={() => {
           setShowBtn(true);
         }}
@@ -209,9 +211,7 @@ function SingleCoreDisplay(props: {
           {props.profile.corrupted ? (
             ""
           ) : (
-            <Box
-              style={{ float: "right", display: "flex", flexDirection: "row" }}
-            >
+            <Box sx={{ float: "right", display: "flex", flexDirection: "row" }}>
               <Fade in={showBtn}>
                 <Box>
                   <Tooltip
@@ -347,6 +347,7 @@ function SingleCoreDisplay(props: {
           onAccept={async () => {
             if (toDestroy) {
               try {
+                setDirtyProfile(props.profile.container, props.profile.id);
                 await remove(
                   getContainer(props.profile.container).getVersionRoot(
                     toDestroy
@@ -376,20 +377,18 @@ function SingleCoreDisplay(props: {
 const PIN_NUMBER_KEY = "PinIndex.";
 
 function getUsed(hash: string): number {
-  return (
-    parseInt(window.localStorage.getItem(PIN_NUMBER_KEY + hash) || "0") || 0
-  );
+  return parseInt(localStorage.getItem(PIN_NUMBER_KEY + hash) || "0") || 0;
 }
 
 function markUsed(hash: string, set?: number): void {
   if (set !== undefined) {
-    window.localStorage.setItem(PIN_NUMBER_KEY + hash, set.toString());
+    localStorage.setItem(PIN_NUMBER_KEY + hash, set.toString());
     return;
   }
   let origin =
-    parseInt(window.localStorage.getItem(PIN_NUMBER_KEY + hash) || "0") || 0;
+    parseInt(localStorage.getItem(PIN_NUMBER_KEY + hash) || "0") || 0;
   origin++;
-  window.localStorage.setItem(PIN_NUMBER_KEY + hash, origin.toString());
+  localStorage.setItem(PIN_NUMBER_KEY + hash, origin.toString());
 }
 
 function getDescriptionFor(type: string): string {
@@ -409,7 +408,7 @@ function CorruptedCoreWarning(): JSX.Element {
   return (
     <>
       <Typography
-        style={{
+        sx={{
           color: "#ff8400",
         }}
         className={"smtxt"}
