@@ -1,18 +1,17 @@
 import {
   Box,
   Button,
-  createStyles,
   FormControlLabel,
-  makeStyles,
-  MuiThemeProvider,
   Radio,
   RadioGroup,
   Switch,
   Tab,
   Tabs,
   TextField,
+  ThemeProvider,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
+import { makeStyles } from "@mui/styles";
 import os from "os";
 import React, { useEffect, useRef, useState } from "react";
 import { DOH_CONFIGURE } from "../modules/commons/Constants";
@@ -28,7 +27,7 @@ import {
 import { loadMirror } from "../modules/download/Mirror";
 import { remoteSelectDir } from "./ContainerManager";
 import { ALICORN_DEFAULT_THEME_LIGHT } from "./Renderer";
-import { useInputStyles } from "./Stylex";
+import { AlicornTheme, useInputStyles } from "./Stylex";
 import { AL_THEMES } from "./ThemeColors";
 import { ALL_ASSISTANTS, tr } from "./Translator";
 
@@ -42,25 +41,24 @@ export enum ConfigType {
 
 export function OptionsPage(): JSX.Element {
   const [tabValue, setTabValue] = useState(0);
-  const classes = makeStyles((theme) =>
-    createStyles({
-      root: {},
-      head: {
-        fontSize: window.sessionStorage.getItem("smallFontSize") || "1em",
-        color: theme.palette.secondary.main,
-      },
-    })
-  )();
+  const classes = makeStyles((theme: AlicornTheme) => ({
+    root: {},
+    head: {
+      fontSize: sessionStorage.getItem("smallFontSize") || "1em",
+      color: theme.palette.secondary.main,
+    },
+  }))();
 
   return (
     <Box className={classes.root}>
-      <MuiThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
+      <ThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
         <Typography className={classes.head}>
           {tr("Options.AutoSave")}
         </Typography>
         <Typography className={classes.head}>{tr("Options.Hint")}</Typography>
         {/* Tabs */}
         <Tabs
+          variant={"fullWidth"}
           value={tabValue}
           onChange={(_e, v) => {
             setTabValue(v);
@@ -129,6 +127,7 @@ export function OptionsPage(): JSX.Element {
             bindConfig={"theme"}
             choices={["Default"].concat(Object.keys(AL_THEMES))}
           />
+          <InputItem type={ConfigType.BOOL} bindConfig={"alicorn-ping"} />
           <InputItem
             reload
             type={ConfigType.STR}
@@ -172,7 +171,13 @@ export function OptionsPage(): JSX.Element {
         </TabPanel>
         <TabPanel index={2} value={tabValue}>
           {/* AL Features */}
+          <InputItem type={ConfigType.BOOL} bindConfig={"readyboom"} />
+          <InputItem type={ConfigType.NUM} bindConfig={"readyboom.cores"} />
           <InputItem type={ConfigType.STR} bindConfig={"hoofoff.central"} />
+          <InputItem
+            type={ConfigType.BOOL}
+            bindConfig={"system.ws-operation"}
+          />
           <InputItem type={ConfigType.BOOL} bindConfig={"command"} />
           <InputItem type={ConfigType.DIR} bindConfig={"cx.shared-root"} />
           <InputItem
@@ -319,17 +324,17 @@ export function OptionsPage(): JSX.Element {
           <InputItem type={ConfigType.BOOL} bindConfig={"reset"} />
           <InputItem type={ConfigType.BOOL} bindConfig={"clean-storage"} />
         </TabPanel>
-      </MuiThemeProvider>
+      </ThemeProvider>
     </Box>
   );
 }
 
 export function hasEdited(conf: string): boolean {
-  return window.localStorage.getItem("Edited." + conf) === "1";
+  return localStorage.getItem("Edited." + conf) === "1";
 }
 
 export function markEdited(conf: string): void {
-  window.localStorage.setItem("Edited." + conf, "1");
+  localStorage.setItem("Edited." + conf, "1");
 }
 
 export function InputItem(props: {
@@ -348,7 +353,7 @@ export function InputItem(props: {
       get(props.bindConfig, undefined) !== originVal.current &&
       props.reload
     ) {
-      window.sessionStorage.setItem("Options.Reload", "1");
+      sessionStorage.setItem("Options.Reload", "1");
     }
     if (props.onChange) {
       props.onChange();
@@ -375,28 +380,26 @@ export function InputItem(props: {
       disabled = true;
     }
   }
-  const classes = makeStyles((theme) =>
-    createStyles({
-      desc: {
-        fontSize: window.sessionStorage.getItem("smallFontSize") || "1em",
-        color: theme.palette.secondary.main,
-      },
-      switch: {
-        color: theme.palette.primary.main,
-        marginLeft: theme.spacing(0.5),
-      },
+  const classes = makeStyles((theme: AlicornTheme) => ({
+    desc: {
+      fontSize: sessionStorage.getItem("smallFontSize") || "1em",
+      color: theme.palette.secondary.main,
+    },
+    switch: {
+      color: theme.palette.primary.main,
+      marginLeft: theme.spacing(0.5),
+    },
 
-      textField: {
-        borderColor: theme.palette.primary.main,
-        color: theme.palette.primary.main,
-      },
-      title: {
-        marginTop: theme.spacing(1),
-        color: theme.palette.primary.main,
-        fontSize: "large",
-      },
-    })
-  )();
+    textField: {
+      borderColor: theme.palette.primary.main,
+      color: theme.palette.primary.main,
+    },
+    title: {
+      marginTop: theme.spacing(1),
+      color: theme.palette.primary.main,
+      fontSize: "large",
+    },
+  }))();
   if (props.experimental) {
     if (!getBoolean("dev.experimental")) {
       return <></>;
@@ -485,7 +488,7 @@ export function InputItem(props: {
                   className={classex.inputDark}
                   type={"button"}
                   variant={"outlined"}
-                  style={{
+                  sx={{
                     marginTop: "0.25em",
                   }}
                   onClick={async () => {
@@ -523,14 +526,7 @@ export function InputItem(props: {
                         <Radio disabled={disabled} checked={cSelect === c} />
                       }
                       label={
-                        <Typography
-                          style={{
-                            color:
-                              ALICORN_DEFAULT_THEME_LIGHT.palette.secondary
-                                .main,
-                          }}
-                          className={"smtxt"}
-                        >
+                        <Typography color={"secondary"} className={"smtxt"}>
                           {tr(`Options.${props.bindConfig}.${c}`)}
                         </Typography>
                       }
