@@ -1,4 +1,4 @@
-import { Typography } from "@material-ui/core";
+import { Typography } from "@mui/material";
 import React from "react";
 import pkg from "../../package.json";
 import { useTextStyles } from "./Stylex";
