@@ -2,24 +2,20 @@ import {
   Box,
   Button,
   FormControl,
-  MuiThemeProvider,
   Tab,
   Tabs,
   TextField,
+  ThemeProvider,
   Typography,
-} from "@material-ui/core";
-import { ipcRenderer } from "electron";
-import React, { useEffect, useRef, useState } from "react";
+} from "@mui/material";
+import React, { useState } from "react";
 import { getBoolean, getString } from "../../modules/config/ConfigSupport";
 import { killEdge, runEdge } from "../../modules/cutie/BootEdge";
 import { applyCode, OnlineGameInfo } from "../../modules/cutie/Hoofoff";
 import { generateWorldAnyUniqueId } from "../../modules/security/Unique";
 import { jumpTo, setChangePageWarn, triggerSetPage } from "../GoTo";
 import { submitInfo, submitSucc, submitWarn } from "../Message";
-import {
-  ALICORN_DEFAULT_THEME_DARK,
-  ALICORN_DEFAULT_THEME_LIGHT,
-} from "../Renderer";
+import { ALICORN_DEFAULT_THEME_LIGHT } from "../Renderer";
 import { useTextStyles } from "../Stylex";
 import { randsl, tr } from "../Translator";
 
@@ -36,20 +32,18 @@ const INTERNET = "internet";
 export function CutieConnet(): JSX.Element {
   const randip = generateRandIP();
   const [superNode, setSuperNode] = useState(
-    window.localStorage.getItem(SUPERNODE_KEY) || ""
+    localStorage.getItem(SUPERNODE_KEY) || ""
   );
-  const [hostIp, setHostIp] = useState(
-    window.localStorage.getItem(IP_KEY) || randip
-  );
+  const [hostIp, setHostIp] = useState(localStorage.getItem(IP_KEY) || randip);
   const [communityName, setCommunityName] = useState(
-    window.localStorage.getItem(COMMUNITY_KEY) ||
+    localStorage.getItem(COMMUNITY_KEY) ||
       randsl("Utilities.CutieConnect.AvailablePrefix") +
         generateWorldAnyUniqueId().slice(-10) // Random name
   );
 
   const [superNodeError, setSuperNodeError] = useState(!checkHost(superNode));
   const [password, setPassword] = useState(
-    window.localStorage.getItem(PASSWORD_KEY) ||
+    localStorage.getItem(PASSWORD_KEY) ||
       randsl("Utilities.CutieConnect.AvailablePassword") +
         generateWorldAnyUniqueId().slice(-20)
   );
@@ -61,6 +55,7 @@ export function CutieConnet(): JSX.Element {
   return (
     <>
       <Tabs
+        variant={"fullWidth"}
         value={tabIndex}
         onChange={(_e, v) => {
           setTabIndex(v);
@@ -74,7 +69,7 @@ export function CutieConnet(): JSX.Element {
           }
         />
         <Tab
-          style={{
+          sx={{
             display: getBoolean("dev.experimental") ? "inherit" : "none",
           }}
           label={
@@ -85,10 +80,10 @@ export function CutieConnet(): JSX.Element {
         />
       </Tabs>
       <TabPanel index={0} value={tabIndex}>
-        <MuiThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
+        <ThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
           <br />
           <br />
-          <FormControl>
+          <FormControl sx={{ width: "100%" }}>
             <TextField
               autoFocus
               color={"primary"}
@@ -153,7 +148,7 @@ export function CutieConnet(): JSX.Element {
               {tr("Utilities.CutieConnect.Disconnect")}
             </Button>
           </FormControl>
-        </MuiThemeProvider>
+        </ThemeProvider>
         <br />
         <br />
         {gameMeta ? (
@@ -180,19 +175,19 @@ export function CutieConnet(): JSX.Element {
           <Typography className={text.secondText}>
             {tr("Utilities.CutieConnect.QuickHint", `RandIP=${randip}`)}
           </Typography>
-          <MuiThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
+          <ThemeProvider theme={ALICORN_DEFAULT_THEME_LIGHT}>
             <TextField
               error={ipError}
               color={"primary"}
               variant={"outlined"}
-              style={{
+              sx={{
                 width: "49%",
                 float: "left",
               }}
               onChange={(e) => {
                 setHostIp(e.target.value.trim());
                 setIPError(!validateIP(e.target.value.trim()));
-                window.localStorage.setItem(IP_KEY, e.target.value.trim());
+                localStorage.setItem(IP_KEY, e.target.value.trim());
               }}
               spellCheck={false}
               margin={"dense"}
@@ -207,16 +202,13 @@ export function CutieConnet(): JSX.Element {
               color={"primary"}
               variant={"outlined"}
               disabled={communityName === INTERNET}
-              style={{
+              sx={{
                 width: "49%",
                 float: "right",
               }}
               onChange={(e) => {
                 setPassword(e.target.value.trim());
-                window.localStorage.setItem(
-                  PASSWORD_KEY,
-                  e.target.value.trim()
-                );
+                localStorage.setItem(PASSWORD_KEY, e.target.value.trim());
               }}
               spellCheck={false}
               margin={"dense"}
@@ -236,12 +228,9 @@ export function CutieConnet(): JSX.Element {
               onChange={(e) => {
                 setSuperNode(e.target.value.trim());
                 setSuperNodeError(!checkHost(e.target.value.trim()));
-                window.localStorage.setItem(
-                  SUPERNODE_KEY,
-                  e.target.value.trim()
-                );
+                localStorage.setItem(SUPERNODE_KEY, e.target.value.trim());
               }}
-              style={{
+              sx={{
                 width: "49%",
                 float: "left",
               }}
@@ -260,12 +249,9 @@ export function CutieConnet(): JSX.Element {
               fullWidth
               onChange={(e) => {
                 setCommunityName(e.target.value.trim());
-                window.localStorage.setItem(
-                  COMMUNITY_KEY,
-                  e.target.value.trim()
-                );
+                localStorage.setItem(COMMUNITY_KEY, e.target.value.trim());
               }}
-              style={{
+              sx={{
                 width: "49%",
                 float: "right",
               }}
@@ -289,10 +275,10 @@ export function CutieConnet(): JSX.Element {
               color={"primary"}
               variant={"contained"}
               onClick={async () => {
-                window.localStorage.setItem(SUPERNODE_KEY, superNode);
-                window.localStorage.setItem(PASSWORD_KEY, password);
-                window.localStorage.setItem(COMMUNITY_KEY, communityName);
-                window.localStorage.setItem(IP_KEY, hostIp); // Freeze Data
+                localStorage.setItem(SUPERNODE_KEY, superNode);
+                localStorage.setItem(PASSWORD_KEY, password);
+                localStorage.setItem(COMMUNITY_KEY, communityName);
+                localStorage.setItem(IP_KEY, hostIp); // Freeze Data
                 await killEdge();
                 await runEdge(
                   communityName,
@@ -306,7 +292,7 @@ export function CutieConnet(): JSX.Element {
               {tr("Utilities.CutieConnect.Connect")}
             </Button>
             <Button
-              style={{ marginLeft: "0.25em" }}
+              sx={{ marginLeft: "0.25em" }}
               color={"primary"}
               variant={"contained"}
               onClick={async () => {
@@ -316,7 +302,7 @@ export function CutieConnet(): JSX.Element {
             >
               {tr("Utilities.CutieConnect.Disconnect")}
             </Button>
-          </MuiThemeProvider>
+          </ThemeProvider>
         </>
       </TabPanel>
     </>
@@ -343,14 +329,6 @@ function generateRandIP(): string {
 
     o.push("172", get16to31Num(), get255Num(), get255Num());
   }
-  return o.join(".");
-}
-
-function generateRandIP10(): string {
-  const o = [];
-  o.push("10", "16", "32");
-  const ds = new Date().getTime();
-  o.push((ds - Math.floor(ds / 256) * 256).toString()); // FIXME: This duplicates fast!
   return o.join(".");
 }
 
@@ -393,68 +371,6 @@ function checkHost(h: string): boolean {
   return false;
 }
 
-function GameDisplay(props: {
-  name: string;
-  desc: string;
-  host: string;
-}): JSX.Element {
-  const text = useTextStyles();
-  const [reachable, setReachable] = useState();
-  const mounted = useRef(false);
-  useEffect(() => {
-    mounted.current = true;
-    return () => {
-      mounted.current = false;
-    };
-  });
-  useEffect(() => {
-    void (async () => {
-      const st = await ipcRenderer.invoke("isReachable", props.host);
-      if (mounted.current) {
-        setReachable(st);
-      }
-    })();
-  }, []);
-  return (
-    <Box
-      style={{
-        marginLeft: "4%",
-      }}
-      onClick={() => {
-        jumpTo("/LaunchPad/" + encodeURIComponent(props.host));
-        triggerSetPage("LaunchPad");
-      }}
-    >
-      <Typography
-        className={text.thirdTextRaw}
-        style={{
-          color:
-            reachable === undefined
-              ? ALICORN_DEFAULT_THEME_DARK.palette.primary.light
-              : reachable
-              ? ALICORN_DEFAULT_THEME_DARK.palette.primary.main
-              : "gray",
-        }}
-      >
-        {props.name}
-      </Typography>
-      <Typography className={text.secondText} gutterBottom>
-        {props.host + " - " + props.desc}
-      </Typography>
-    </Box>
-  );
-}
-
-function getMapHostBySupernode(h: string): string {
-  const ps = h.split(":");
-  ps[1] = (parseInt(ps[1]) - 1).toString();
-  return "ws://" + ps.join(":");
-}
-
-function validatePort(p: string): boolean {
-  const s = parseInt(p);
-  return s >= 0 && s <= 65535;
-}
 function TabPanel(props: {
   children?: React.ReactNode;
   index: string | number;
