@@ -1,3 +1,5 @@
+import path from "path";
+
 export abstract class AbstractDownloader {
   abstract downloadFile(meta: DownloadMeta): Promise<DownloadStatus>;
 }
@@ -9,7 +11,9 @@ export class DownloadMeta {
   readonly size: number;
   constructor(url: string, savePath: string, sha1 = "", size = 0) {
     this.url = url;
-    this.savePath = savePath;
+    this.savePath = path.isAbsolute(savePath)
+      ? path.normalize(savePath)
+      : path.resolve(savePath);
     this.sha1 = sha1;
     this.size = size;
   }
